MINTEMP, Slightly lower temperature before triggering sensor fault.
In Scotland it gets cold.
At startup in Winter, the temps were reading 3-4 degrees, so triggering a sensor missing fault.  Until warmed up with a hairdryer !
I'm not sure if zero, or negative value is acceptable so I've changed MINTEMP from 5 to 1.
If zero is OK then I'd go for that.